# backend/docs/openapi.json
{
    "components": {
        "schemas": {
            "admin.AuditListResponse": {
                "properties": {
                    "limit": {
                        "example": 20,
                        "type": "integer"
                    },
                    "logs": {
                        "items": {
                            "$ref": "#/components/schemas/admin.AuditLogResponse"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "page": {
                        "example": 1,
                        "type": "integer"
                    },
                    "total": {
                        "example": 15,
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "admin.SessionTokenRequest": {
                "properties": {
                    "token": {
                        "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "admin.UserMessageResponse": {
                "properties": {
                    "message": {
                        "example": "Rol actualizado correctamente",
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/auth.UserResponse"
                    }
                },
                "type": "object"
            },
...
            "utils.ErrorResponse": {
                "properties": {
                    "error": {
                        "example": "Se requieren username, email y password",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "utils.MessageResponse": {
                "properties": {
                    "message": {
                        "example": "Operación exitosa",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "description": "Token JWT de sesión. Formato: `Bearer \u003ctoken\u003e`.",
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "description": "API REST del monolito modular: identidad y sesiones, administración y auditoría, autoría de cursos versionados y carga multimedia con procesamiento asíncrono a HLS.\nAutenticación con JWT revocable: iniciar sesión en /api/v1/auth/login y enviar `Authorization: Bearer \u003ctoken\u003e`.",
        "title": "Plataforma MOOC - API",
        "version": "1.0"
    },
...
    "paths": {
        "/api/v1/admin/audit": {
            "get": {
                "description": "Registros inmutables de acciones administrativas, filtrables por actor, objetivo, acción y rango de fechas.",
                "parameters": [
                    {
                        "description": "Bearer \u003ctoken de admin\u003e",
                        "in": "header",
                        "name": "Authorization",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filtro por actor (username)",
                        "in": "query",
                        "name": "actor",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filtro por usuario objetivo",
                        "in": "query",
                        "name": "target",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filtro por acción",
                        "in": "query",
                        "name": "action",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Desde (fecha)",
                        "in": "query",
                        "name": "from",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Hasta (fecha)",
                        "in": "query",
                        "name": "to",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Página (base 1)",
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "default": 1,
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Resultados por página (máx 100)",
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "default": 20,
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/admin.AuditListResponse"
                                }
                            }
                        },
                        "description": "Registros de auditoría"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "Error interno"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Listar auditoría",
                "tags": [
                    "Administración"
                ]
            }
        },
        "/api/v1/admin/sessions": {
            "delete": {
                "description": "Revoca el token indicado (cuerpo o query) con registro en auditoría.",
                "parameters": [
                    {
                        "description": "Bearer \u003ctoken de admin\u003e",
                        "in": "header",
                        "name": "Authorization",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Token a revocar (alternativo)",
                        "in": "query",
                        "name": "token",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/admin.SessionTokenRequest",
                                "summary": "request",
                                "description": "Token a revocar (o query param token)"
                            }
                        }
                    },
                    "description": "Token a revocar (o query param token)"
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.MessageResponse"
                                }
                            }
                        },
                        "description": "Sesión revocada"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "Falta el token a revocar"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "Error interno"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Revocar una sesión",
                "tags": [
                    "Administración"
                ]
            }
        },
        "/api/v1/admin/stats": {
            "get": {
                "description": "Totales de usuarios por estado y administradores activos.",
                "parameters": [
                    {
                        "description": "Bearer \u003ctoken de admin\u003e",
                        "in": "header",
                        "name": "Authorization",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/admin.StatsResponse"
                                }
                            }
                        },
                        "description": "Estadísticas"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Estadísticas",
                "tags": [
                    "Administración"
                ]
            }
        },
        "/api/v1/admin/users": {
            "get": {
                "description": "Lista paginada de usuarios con búsqueda por texto, filtros por rol, estado y verificación.",
                "parameters": [
                    {
                        "description": "Bearer \u003ctoken de admin\u003e",
                        "in": "header",
                        "name": "Authorization",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Búsqueda en username y email",
                        "in": "query",
                        "name": "search",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filtro por rol (student, professor, admin)",
                        "in": "query",
                        "name": "role",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filtro por estado (active, inactive, blocked)",
                        "in": "query",
                        "name": "status",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filtro por verificación",
                        "in": "query",
                        "name": "verified",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "description": "Página (base 1)",
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "default": 1,
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Resultados por página (máx 100)",
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "default": 20,
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Orden: columna y dirección (ej. username asc)",
                        "in": "query",
                        "name": "sort",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/admin.UserListResponse"
                                }
                            }
                        },
                        "description": "Listado de usuarios"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "No autenticado"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "Se requiere rol admin"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "Error interno"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Listar usuarios",
                "tags": [
                    "Administración"
                ]
            }
        },
...
        "/api/v1/admin/users/{id}/restore": {
            "post": {
                "description": "Revierte el soft delete de un usuario, con registro en auditoría.",
                "parameters": [
                    {
                        "description": "Bearer \u003ctoken de admin\u003e",
                        "in": "header",
                        "name": "Authorization",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "ID del usuario",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/admin.UserMessageResponse"
                                }
                            }
                        },
                        "description": "Usuario restaurado"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "ID inválido"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "Error interno"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Restaurar usuario",
                "tags": [
                    "Administración"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "localhost:8080"
        }
    ]
}